Push notification fix Details: - Updated the push notification registration logic. - Updated logic to ask permission for push notification.
